time bot out and freeze them, some groundwork for freezing crashed bots
